Tidy up styling and move methods into component directories with better naming
